common: Move video config box parsers to common
